Linux监控脚本/requirements.txt
PyYAML==6.0.1
requests==2.28.1
